Move FC_Board contents to repo root (#76)
