+++ lib/bitflags.py
