+++ lib/debugcolor.py
