+++ lib/detumble.py
