+++ lib/functions.py
