+++ safemode.py
