+++ tests/repl/echo.py
