+++ tests/unit/test_detumble.py
@@ -5,8 +5,7 @@
 
 import pytest
 
-# from Batt_Board import detumble
-import detumble
+import lib.detumble as detumble
 
 
 def test_dot_product():
